requirements.txt
ipython==8.0.1
discord.py==1.7.3

